requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
MarkupSafe==2.0.1
Flask-RESTful
mysql-connector-python
psycopg2-binary
passlib
Flask-JWT-Extended
email-validator
requests
chardet

